# apps/web-admin/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "仓脉 WMS API",
    "description": "仓脉 WMS HTTP API（OpenAPI 3，FastAPI 自动生成）。\n\n## 在 Swagger UI 中试调\n\n1. 调用 `POST /api/v1/auth/login`，使用开发种子账号拿到 `access_token`\n2. 点击右上角 **Authorize**，值填入 token（无需手写 `Bearer ` 前缀；若 UI 要求完整头则用 `Bearer <token>`）\n3. 再调用受保护接口；入库上架、出库审核/拣货/取消、盘点创建/审核/取消等写操作还需请求头 `Idempotency-Key`\n\n生产环境（`APP_ENV=prod`/`production`）不暴露本页、`/redoc` 与 `/openapi.json`。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/inbound-orders": {
      "post": {
        "tags": [
          "inbound"
        ],
        "summary": "Create Inbound Order",
        "operationId": "create_inbound_order_api_v1_inbound_orders_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboundOrderCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "inbound"
        ],
        "summary": "List Inbound Orders",
        "operationId": "list_inbound_orders_api_v1_inbound_orders_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Warehouse Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "order_no",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order No"
            }
          },
          {
            "name": "order_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Type"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码，从 1 开始",
              "default": 1,
              "title": "Page"
            },
            "description": "页码，从 1 开始"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页条数，默认 20，上限 100",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页条数，默认 20，上限 100"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Inbound Orders Api V1 Inbound Orders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inbound-orders/{order_id}": {
      "get": {
        "tags": [
          "inbound"
        ],
        "summary": "Get Inbound Order",
        "operationId": "get_inbound_order_api_v1_inbound_orders__order_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Inbound Order Api V1 Inbound Orders  Order Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "inbound"
        ],
        "summary": "Update Inbound Order",
        "operationId": "update_inbound_order_api_v1_inbound_orders__order_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboundOrderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Inbound Order Api V1 Inbound Orders  Order Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inbound-orders/{order_id}/submit": {
      "post": {
        "tags": [
          "inbound"
        ],
        "summary": "Submit Inbound Order",
        "operationId": "submit_inbound_order_api_v1_inbound_orders__order_id__submit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Submit Inbound Order Api V1 Inbound Orders  Order Id  Submit Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inbound-orders/{order_id}/approve": {
      "post": {
        "tags": [
          "inbound"
        ],
        "summary": "Approve Inbound Order",
        "operationId": "approve_inbound_order_api_v1_inbound_orders__order_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Approve Inbound Order Api V1 Inbound Orders  Order Id  Approve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inbound-orders/{order_id}/cancel": {
      "post": {
        "tags": [
          "inbound"
        ],
        "summary": "Cancel Inbound Order",
        "operationId": "cancel_inbound_order_api_v1_inbound_orders__order_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Inbound Order Api V1 Inbound Orders  Order Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inbound-orders/{order_id}/putaway": {
      "post": {
        "tags": [
          "inbound"
        ],
        "summary": "Putaway Inbound Order",
        "operationId": "putaway_inbound_order_api_v1_inbound_orders__order_id__putaway_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutawayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Putaway Inbound Order Api V1 Inbound Orders  Order Id  Putaway Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InboundLineInput": {
        "properties": {
          "sku_id": {
            "type": "integer",
            "title": "Sku Id"
          },
          "planned_qty": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Planned Qty"
          }
        },
        "type": "object",
        "required": [
          "sku_id",
          "planned_qty"
        ],
        "title": "InboundLineInput"
      },
      "InboundOrderCreate": {
        "properties": {
          "warehouse_id": {
            "type": "integer",
            "title": "Warehouse Id"
          },
          "order_type": {
            "type": "string",
            "enum": [
              "purchase",
              "return",
              "other"
            ],
            "title": "Order Type"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/InboundLineInput"
            },
            "type": "array",
            "minItems": 1,
            "title": "Lines"
          },
          "supplier_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          },
          "order_no": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Order No"
          }
        },
        "type": "object",
        "required": [
          "warehouse_id",
          "order_type",
          "lines"
        ],
        "title": "InboundOrderCreate"
      },
      "InboundOrderUpdate": {
        "properties": {
          "lines": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/InboundLineInput"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lines"
          },
          "supplier_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          }
        },
        "type": "object",
        "title": "InboundOrderUpdate"
      },
...
      "PutawayRequest": {
        "properties": {
          "line_id": {
            "type": "integer",
            "title": "Line Id"
          },
          "location_id": {
            "type": "integer",
            "title": "Location Id"
          },
          "qty": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Qty"
          }
        },
        "type": "object",
        "required": [
          "line_id",
          "location_id",
          "qty"
        ],
        "title": "PutawayRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}